split out a few remaining shaping checks
